Fix bug parsing large streaming JSON objects 🐽
Internally, JsonStreamParser uses a buffered reader to read the streaming HTTP response. It is possible for a single chunk to exceed the size of this buffer. Previously, if one of these reads resulted in an empty chunk this could cause the stream to get “stuck” perpetually trying to parse an incomplete object. When this condition was triggered the k8s namer would stop processing updates for an endpoint until the process was restarted.

I’ve also refactored the code, hopefully making it a lot simpler.