Fix bug parsing large streaming JSON objects 🐽
Internally, JsonStreamParser uses a buffered reader to read the streaming HTTP response. It is possible for a single chunk to exceed the size of this buffer. Previously, this could cause the stream to get “stuck” perpetually trying to parse an incomplete object.

I’ve also refactored the code, hopefully making it a lot simpler.